requirements.txt
google-generativeai
python-dotenv
pandas
numpy
beautifulsoup4
chromadb
uvicorn 
fastapi 
